refactor: extract classes from cli.py
